Module Belt.Id

Belt.Id

Provide utilities to create identified comparators or hashes for data structures used below.

It create a unique identifier per module of functions so that different data structures with slightly different comparison functions won't mix

type ('a, 'id) hash = 'a -> int

type ('a, 'id) eq = 'a -> 'a -> bool

type ('a, 'id) cmp = 'a -> 'a -> int

val getHashInternal : ('a, 'id) hash -> 'a -> int

val getEqInternal : ('a, 'id) eq -> 'a -> 'a -> bool

val getCmpInternal : ('a, 'id) cmp -> 'a -> 'a -> int

module type Comparable = sig ... end

type ('key, 'id) comparable = (module Comparable with type identity = 'id and type t = 'key)

module MakeComparableU (M : sig ... end) : sig ... end

module MakeComparable (M : sig ... end) : sig ... end

val comparableU : cmp:('key -> 'key -> int) -> (module Comparable with type t = 'key)

val comparable : cmp:('key -> 'key -> int) -> (module Comparable with type t = 'key)

module type Hashable = sig ... end

type ('key, 'id) hashable = (module Hashable with type identity = 'id and type t = 'key)

module MakeHashableU (M : sig ... end) : sig ... end

module MakeHashable (M : sig ... end) : sig ... end

val hashableU : hash:('key -> int) -> eq:('key -> 'key -> bool) -> (module Hashable with type t = 'key)

val hashable : hash:('key -> int) -> eq:('key -> 'key -> bool) -> (module Hashable with type t = 'key)